requirements.txt
Pillow==10.2.0
streamlit==1.32.2
torch==2.2.1
transformers==4.40.0.dev0
yolov5==7.0.13
